Translate label and migrate data with SQL
With the built-in methods, each record is changed one by one.
SQL allows for batch-processing. This comes at the cost that
custom SQL needed to be written. The speed-boost, however, is
noticable. The AR-version didn't even respond for several minutes,
the SQL-version completes within 20 seconds locally and gives
feedback on the progress.